+++ src/mol-gl/shader/cylinders.frag.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020-2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2020-2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  * @author Gianluca Tomasello <giagitom@gmail.com>
@@ -35,6 +35,43 @@ uniform vec4 uInteriorSubstance;
     const bool solidInterior = true;
 #else
     const bool solidInterior = false;
+#endif
+
+bool isCap = false;
+
+#ifdef dSolidInterior
+    bool solidInteriorCap(
+        in vec3 rayOrigin, in vec3 rayDir, in float frontT, in float t,
+        in float frontDepth, in bool objectClipped, in vec3 frontModelPosition,
+        inout vec3 cameraNormal, inout vec3 modelPosition, inout vec3 viewPosition, inout float fragmentDepth
+    ){
+        vec3 cameraRayOrigin = (uView * vec4(rayOrigin, 1.0)).xyz;
+        vec3 cameraRayDir = (uView * vec4(rayDir, 0.0)).xyz;
+        float nearT = - (uNear + cameraRayOrigin.z) / cameraRayDir.z;
+        float sNear = frontDepth > 0.0 ? 0.0 : clamp((nearT - frontT) / (t - frontT), 0.0, 1.0);
+        #if defined(dClipVariant_pixel) && dClipObjectCount != 0
+            float sCap = objectClipped || frontDepth <= 0.0 ? clipExit(frontModelPosition / uModelScale, modelPosition / uModelScale, sNear) : sNear;
+            if (sCap < 0.0) return false;
+        #else
+            float sCap = sNear;
+        #endif
+        if (sCap == sNear && frontDepth <= 0.0) {
+            fragmentDepth = 0.0 + (0.0000002 / vSize);
+            cameraNormal = -rayDir;
+            viewPosition = cameraRayOrigin + nearT * cameraRayDir;
+            modelPosition = (uInvView * vec4(viewPosition, 1.0)).xyz;
+        }
+        #if defined(dClipVariant_pixel) && dClipObjectCount != 0
+            else if (objectClipped || frontDepth <= 0.0) {
+                modelPosition = mix(frontModelPosition, modelPosition, sCap);
+                viewPosition = (uView * vec4(modelPosition, 1.0)).xyz;
+                fragmentDepth = calcDepth(viewPosition) + (0.0000002 / vSize);
+                cameraNormal = -clipNormal(modelPosition / uModelScale);
+                isCap = true;
+            }
+        #endif
+        return true;
+    }
 #endif
 
 // adapted from https://www.shadertoy.com/view/4lcSRn
@@ -74,6 +111,9 @@ bool CylinderImpostor(
 
     bool clipped = false;
     bool objectClipped = false;
+    float frontT = 0.0;
+    float frontDepth = 1.0;
+    vec3 frontModelPosition = rayOrigin;
 
     #if defined(dRenderVariant_depth)
         // back-depth pass wants the farthest intersection
@@ -92,8 +132,11 @@ bool CylinderImpostor(
         modelPosition = rayOrigin + t * rayDir;
         viewPosition = (uView * vec4(modelPosition, 1.0)).xyz;
         fragmentDepth = calcDepth(viewPosition);
+        frontT = t;
+        frontDepth = fragmentDepth;
+        frontModelPosition = modelPosition;
         #if defined(dClipVariant_pixel) && dClipObjectCount != 0
-            if (clipTest(modelPosition)) {
+            if (clipTest(modelPosition / uModelScale)) {
                 objectClipped = true;
                 fragmentDepth = -1.0;
                 #ifdef dSolidInterior
@@ -116,8 +159,11 @@ bool CylinderImpostor(
                 modelPosition = rayOrigin + t * rayDir;
                 viewPosition = (uView * vec4(modelPosition, 1.0)).xyz;
                 fragmentDepth = calcDepth(viewPosition);
+                frontT = t;
+                frontDepth = fragmentDepth;
+                frontModelPosition = modelPosition;
                 #if defined(dClipVariant_pixel) && dClipObjectCount != 0
-                    if (clipTest(modelPosition)) {
+                    if (clipTest(modelPosition / uModelScale)) {
                         objectClipped = true;
                         fragmentDepth = -1.0;
                         #ifdef dSolidInterior
@@ -146,8 +192,11 @@ bool CylinderImpostor(
                 modelPosition = rayOrigin + t * rayDir;
                 viewPosition = (uView * vec4(modelPosition, 1.0)).xyz;
                 fragmentDepth = calcDepth(viewPosition);
+                frontT = t;
+                frontDepth = fragmentDepth;
+                frontModelPosition = modelPosition;
                 #if defined(dClipVariant_pixel) && dClipObjectCount != 0
-                    if (clipTest(modelPosition)) {
+                    if (clipTest(modelPosition / uModelScale)) {
                         objectClipped = true;
                         fragmentDepth = -1.0;
                         #ifdef dSolidInterior
@@ -183,17 +232,7 @@ bool CylinderImpostor(
             fragmentDepth = calcDepth(viewPosition);
             if (fragmentDepth > 0.0) {
                 #ifdef dSolidInterior
-                    if (!objectClipped && !depthBack) {
-                        fragmentDepth = 0.0 + (0.0000002 / vSize);
-                        cameraNormal = -rayDir;
-
-                        // intersection of ray in model space with near plane in camera space
-                        vec3 cameraRayOrigin = (uView * vec4(rayOrigin, 1.0)).xyz;
-                        vec3 cameraRayDir = (uView * vec4(rayDir, 0.0)).xyz;
-                        float nearT = - (uNear + cameraRayOrigin.z) / cameraRayDir.z;
-                        viewPosition = cameraRayOrigin + nearT * cameraRayDir;
-                        modelPosition = (uInvView * vec4(viewPosition, 1.0)).xyz;
-                    }
+                    if (!depthBack && !solidInteriorCap(rayOrigin, rayDir, frontT, t, frontDepth, objectClipped, frontModelPosition, cameraNormal, modelPosition, viewPosition, fragmentDepth)) return false;
                 #endif
                 return true;
             }
@@ -210,17 +249,7 @@ bool CylinderImpostor(
                 fragmentDepth = calcDepth(viewPosition);
                 if (fragmentDepth > 0.0) {
                     #ifdef dSolidInterior
-                        if (!objectClipped && !depthBack) {
-                            fragmentDepth = 0.0 + (0.0000002 / vSize);
-                            cameraNormal = -rayDir;
-
-                            // intersection of ray in model space with near plane in camera space
-                            vec3 cameraRayOrigin = (uView * vec4(rayOrigin, 1.0)).xyz;
-                            vec3 cameraRayDir = (uView * vec4(rayDir, 0.0)).xyz;
-                            float nearT = - (uNear + cameraRayOrigin.z) / cameraRayDir.z;
-                            viewPosition = cameraRayOrigin + nearT * cameraRayDir;
-                            modelPosition = (uInvView * vec4(viewPosition, 1.0)).xyz;
-                        }
+                        if (!depthBack && !solidInteriorCap(rayOrigin, rayDir, frontT, t, frontDepth, objectClipped, frontModelPosition, cameraNormal, modelPosition, viewPosition, fragmentDepth)) return false;
                     #endif
                     return true;
                 }
@@ -236,17 +265,7 @@ bool CylinderImpostor(
                 fragmentDepth = calcDepth(viewPosition);
                 if (fragmentDepth > 0.0) {
                     #ifdef dSolidInterior
-                        if (!objectClipped && !depthBack) {
-                            fragmentDepth = 0.0 + (0.0000002 / vSize);
-                            cameraNormal = -rayDir;
-
-                            // intersection of ray in model space with near plane in camera space
-                            vec3 cameraRayOrigin = (uView * vec4(rayOrigin, 1.0)).xyz;
-                            vec3 cameraRayDir = (uView * vec4(rayDir, 0.0)).xyz;
-                            float nearT = - (uNear + cameraRayOrigin.z) / cameraRayDir.z;
-                            viewPosition = cameraRayOrigin + nearT * cameraRayDir;
-                            modelPosition = (uInvView * vec4(viewPosition, 1.0)).xyz;
-                        }
+                        if (!depthBack && !solidInteriorCap(rayOrigin, rayDir, frontT, t, frontDepth, objectClipped, frontModelPosition, cameraNormal, modelPosition, viewPosition, fragmentDepth)) return false;
                     #endif
                     return true;
                 }
@@ -277,7 +296,9 @@ void main() {
     vec3 vModelPosition = modelPosition;
 
     #include fade_lod
-    #include clip_pixel
+    if (!isCap) {
+        #include clip_pixel
+    }
 
     #ifdef dNeedsNormal
         mat3 normalMatrix = adjoint(uView);